Code: check if user has created a course

+++ src/services/DeleteUserService.ts
@@ -1,6 +1,7 @@
 import { getRepository } from 'typeorm'
 import Enrollment from '../models/Enrollment'
 import User from '../models/User'
+import Course from '../models/Course'
 import AppError from '../errors/AppError'
 
 class DeleteUserService {
@@ -8,6 +9,7 @@ class DeleteUserService {
   public async execute(id: string): Promise<boolean> {
     const userRepository = getRepository(User)
     const enrollmentsRepository = getRepository(Enrollment)
+    const courseRepository = getRepository(Course)
 
     const userRegistered = await userRepository.findOne({
       where: {
@@ -26,6 +28,16 @@ class DeleteUserService {
 
     if (enrollments)
       throw new AppError('This user is enrolled to a course. Before delete it, delete the enrollment.', 400)
+    
+    const courseRegistered = await courseRepository.findOne({
+      where: {
+        created_by: id
+      }
+    })
+  
+    if (courseRegistered) {
+      throw new AppError('This user has created a course. Before delete it, delete the course.', 400)
+    }
 
     await userRepository.delete(id)
